Create a top-level player_interaction module
