+++ emergence_game/src/main.rs
@@ -16,7 +16,7 @@ fn main() {
         .add_plugin(emergence_lib::simulation::SimulationPlugin {
             gen_config: GenerationConfig::default(),
         })
-        .add_plugin(emergence_lib::InteractionPlugin)
+        .add_plugin(emergence_lib::player_interaction::InteractionPlugin)
         .add_plugin(emergence_lib::graphics::GraphicsPlugin)
         .run();
 }
